Minor updates on validators
