+++ client/src/components/Login/Login.jsx
@@ -1,7 +1,7 @@
-import { useEffect, useState } from "react";
+import { useCallback, useState } from "react";
 import styles from "./Login.module.scss";
-import { emailValidation } from "../../utils/emailValidator";
-import { passwordValidation } from "../../utils/passwordValidator";
+import { emailValidation } from "../../utils/validators";
+import { passwordValidation } from "../../utils/validators";
 
 const Login = () => {
   const [email, setEmail] = useState("");
@@ -17,7 +17,7 @@ const Login = () => {
 
   const [isEmailValid, setIsEmailValid] = useState();
 
-  useEffect(() => {
+  const checkIsMailValid = useCallback(() => {
     if (emailValidation(email)) {
       setIsEmailValid(true);
     } else {
@@ -27,7 +27,7 @@ const Login = () => {
 
   const [isPasswordValid, setIsPasswordValid] = useState();
 
-  useEffect(() => {
+  const checkIspasswordValid = useCallback(() => {
     if (passwordValidation(password)) {
       setIsPasswordValid(true);
     } else {
@@ -52,7 +52,9 @@ const Login = () => {
           placeholder="Email"
           value={email}
           onChange={handleEmail}
+          onBlur={checkIsMailValid}
         />
+        {isEmailValid === false && <span>Please enter a valid email</span>}
       </div>
       <div className={styles.loginInputs}>
         <input
@@ -61,9 +63,18 @@ const Login = () => {
           placeholder="Password"
           value={password}
           onChange={handlePassword}
+          onBlur={checkIspasswordValid}
         />
+        {isPasswordValid === false && (
+          <span>Please enter a valid password</span>
+        )}
       </div>
-      <button onClick={handleLogin}>Login</button>
+      <button
+        onClick={handleLogin}
+        disabled={!isEmailValid || !isPasswordValid}
+      >
+        Login
+      </button>
     </div>
   );
 };
