+++ client/src/utils/validators.js
@@ -7,3 +7,13 @@ export const passwordValidation = (enteredValue) => {
     return true;
   }
 };
+
+export const emailValidation = (enteredValue) => {
+  const regex = /^([^\s\@])+\@(([^\s\@\.])+\.)+([^\s\.]{2,})+$/;
+  const regexResult = regex.test(enteredValue);
+  if (enteredValue === "" || !regexResult) {
+    return false;
+  } else {
+    return true;
+  }
+};
